requirements.txt
python-telegram-bot>=20.0
asyncpg>=0.27.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
